Rename constance to constants
