+++ src/modules/metrics/metrics.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectMetric } from '@willsoto/nestjs-prometheus';
 import { Counter } from 'prom-client';
 
-import { DATA_FROM_API, DATA_FROM_REDIS_CACHE } from './constance/metric-names';
+import { DATA_FROM_API, DATA_FROM_REDIS_CACHE } from './constants/metric-names';
 
 @Injectable()
 export class MetricsService {
